Fixed line 36 in app-docker.py and updated README.md
Reason for changes:

The original line 36 was assigning a tuple to the centerText variable. False should be inside the os.getenv function as the default parameter. Also, because Python sees ENV variables as text they need to be in quotes.

I also added two more edge-case checks.
  - Added ".lower" to the output in case the user uses upper-case so the boolean won't fail.
  - Added "== 'true'" to the end of the line so that the assignment will be false if garbage is input instead of "true" or "false" then it will default to "false".

I also updated the README.md file to reflect the new additions.

Chad